Pull out confirm and unixTimeToDate
