+++ src/Utilities/Input.h
@@ -1,0 +1,12 @@
+// Copyright (c) 2018, The TurtleCoin Developers
+// 
+// Please see the included LICENSE file for more information.
+
+#include <string>
+
+namespace Utilities
+{
+    bool confirm(const std::string &msg);
+
+    bool confirm(const std::string &msg, const bool defaultReturn);
+}
