+++ src/zedwallet++/CommandDispatcher.cpp
@@ -8,8 +8,10 @@
 
 #include <iostream>
 
-#include <zedwallet++/AddressBook.h>
 #include <Utilities/ColouredMsg.h>
+#include <Utilities/Input.h>
+
+#include <zedwallet++/AddressBook.h>
 #include <zedwallet++/CommandImplementations.h>
 #include <zedwallet++/Open.h>
 #include <zedwallet++/Transfer.h>
@@ -105,7 +107,7 @@ bool handleCommand(
                      "send large amounts at once.\n"
                   << WarningMsg("This may take a very long time!\n");
 
-        if (!ZedUtilities::confirm("Do you want to proceed?"))
+        if (!Utilities::confirm("Do you want to proceed?"))
         {
             std::cout << WarningMsg("Cancelling optimization.") << std::endl;
         }
