+++ src/zedwallet++/CommandImplementations.cpp
@@ -16,6 +16,7 @@
 
 #include <Utilities/ColouredMsg.h>
 #include <Utilities/FormatTools.h>
+#include <Utilities/Input.h>
 #include <Utilities/String.h>
 
 #include <zedwallet++/Commands.h>
@@ -300,7 +301,7 @@ void reset(const std::shared_ptr<WalletBackend> walletBackend)
               << InformationMsg("process.")
               << std::endl << std::endl;
     
-    if (!ZedUtilities::confirm("Are you sure?"))
+    if (!Utilities::confirm("Are you sure?"))
     {
         return;
     }
@@ -362,7 +363,7 @@ void saveCSV(const std::shared_ptr<WalletBackend> walletBackend)
 
         const std::string direction = tx.totalAmount() > 0 ? "IN" : "OUT";
 
-        csv << ZedUtilities::unixTimeToDate(tx.timestamp) << ","    /* Timestamp */
+        csv << Utilities::unixTimeToDate(tx.timestamp) << ","    /* Timestamp */
             << tx.blockHeight << ","                                /* Block Height */
             << tx.hash << ","                                       /* Hash */
             << amount << ","                                        /* Amount */
@@ -388,7 +389,7 @@ void printOutgoingTransfer(const WalletTypes::Transaction tx)
     if (tx.blockHeight != 0 && tx.timestamp != 0)
     {
         stream << "Block height: " << tx.blockHeight << "\n"
-               << "Timestamp: " << ZedUtilities::unixTimeToDate(tx.timestamp) << "\n";
+               << "Timestamp: " << Utilities::unixTimeToDate(tx.timestamp) << "\n";
     }
 
     stream << "Spent: " << Utilities::formatAmount(amount - tx.fee) << "\n"
@@ -411,7 +412,7 @@ void printIncomingTransfer(const WalletTypes::Transaction tx)
 
     stream << "Incoming transfer:\nHash: " << tx.hash << "\n"
            << "Block height: " << tx.blockHeight << "\n"
-           << "Timestamp: " << ZedUtilities::unixTimeToDate(tx.timestamp) << "\n"
+           << "Timestamp: " << Utilities::unixTimeToDate(tx.timestamp) << "\n"
            << "Amount: " << Utilities::formatAmount(amount) << "\n";
 
     if (tx.paymentID != "")
@@ -432,7 +433,7 @@ void printIncomingTransfer(const WalletTypes::Transaction tx)
                   << InformationMsg(tx.unlockTime)
                   << std::endl
                   << InformationMsg("Unlocks at approximately: ")
-                  << InformationMsg(ZedUtilities::unixTimeToDate(unlockInUnixTime))
+                  << InformationMsg(Utilities::unixTimeToDate(unlockInUnixTime))
                   << std::endl
                   << std::endl;
     }
@@ -441,7 +442,7 @@ void printIncomingTransfer(const WalletTypes::Transaction tx)
     {
         std::cout << SuccessMsg(stream.str())
                   << InformationMsg("Unlocks at: ")
-                  << InformationMsg(ZedUtilities::unixTimeToDate(tx.unlockTime))
+                  << InformationMsg(Utilities::unixTimeToDate(tx.unlockTime))
                   << std::endl
                   << std::endl;
     }
